pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.26.0",
    "mcp[cli]>=1.11.0",
    "python-dotenv>=1.1.1",
]
